requirements.txt
Flask>=3.1
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-WTF>=1.2
Pillow>=10.0
pillow-avif-plugin>=1.5
python-dotenv>=1.0
